requirements.txt
click==8.1.3
Pillow==9.2.0
pyserial==3.5
Wand==0.6.10
qrcode==7.4.2
matplotlib==3.7.1
pygame==2.3.0
